Fix DeepEquals and ToString
Stupid Robert - deepequals was not doing correct comparision, and also there was a problem with reading two JSONs one with 30.0 and one with 30, they are value the same but end up with different token types.  it now copes with it.

ToString - if its double, always print a .X so it keeps are double when its outputted as a string.